Basic version of trendles' homepage and photography club page.
Photography page only contains functional units. It's design is incomplete yet. Moreover IMO it should only be displaying pictures, uploading them should be the task of the admin. Couldn't think of a design yet so I have added upload facility to that page itself.
Guide me to make it better.